refactor: rename of the http downloader to real debrid downloader given that it only supports downloading games through real debrid
